support optional inquiry-data logging in NetLogo step
If a NetLogo model is used that has integrated the 
inquiry-data logging module developed for the VISUAL
project then this updated NetLogo step type will extract
the inquiry-log data and stave it in the step session.

See this example of an AirBag model that has integrated
the inquiry-data logging module.

https://gist.github.com/stepheneb/5469959

The integrated inquiry-data logging module is derived from
the work shown here developed for the VISUAL project:

http://visual.concord.org/experiments/netlogo-visual/models.html

In these examples the module is loaded separately from
the actual NetLogo model with the following statement in the
very beginning of the NetLogo code for the model:

  __includes [ "data-export-modular.nls" ]

Here's where these resources/examples are available in a code repository:
https://github.com/concord-consortium/lab/tree/WIP-refactor-properties/src/experiments/netlogo-visual/models

However there is a bug in the Windows version of the
NetLogo applet and the only way an external module can
be loaded is if it comes from the same directory location
as the NetLogo *.logo file.

If the data-export-modular.nls module was included as a
resource in this step the web application would need to
synthesize a path for any NetLogo applet that requested
it so it appeared to the NetLogo applet that data-export-modular.nls
was being served from the same directory.

This of course would be a big improvement because with
this initial implementation the code in data-export-modular.nls
needs to be copied into every NetLogo applet that uses it.

If there is a bug fixed in data-export-modular.nls each
NetLogo applet that uses it will need to be updated.